+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - `--free` to `mia model list` that is equivalent to `--max-price 0`.
 - `mia models` is now an alias for `mia model`.
 
+### Fixed
+- `--max-price` not evaluating properly
+
 ## v0.1.1
 ### Added
 - Multiline inputs (Shift+Enter/Alt+Enter) and file pasting in tui.